backend/api/requirements.txt
flask
firebase-admin
openai
python-dotenv